backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
pandas==2.3.2
numpy==2.3.3
scipy==1.16.2
statsmodels==0.14.5
scikit-learn==1.7.2
fairlearn==0.11.0
shap==0.48.0
lime==0.2.0.1
boto3==1.40.30
reportlab==4.4.3
httpx==0.28.1
email-validator==2.3.0
groq>=0.30.0,<1.0.0
google-auth==2.41.1
requests==2.32.5
python-json-logger==2.0.7
pyyaml==6.0.2
aiofiles==23.2.1
openpyxl==3.1.5
